Helper Functions Broken
Fixing helper functions. These were broken becasue of the use of
`blocking_send` for tokio channels. Ultimately the lua functions are
executied within the async tokio context but the compiler has no way to
know this. So you can't use await inside a non-async function but
blocking inside the async context will fail. So you must spawn out a new
async task and await the async send for this to work properly.